+++ klibc/mmap.cpp
@@ -1,6 +1,7 @@
 #include "klibc/mmap.hpp"
 
 #include "kernel/kprint.hpp"
+#include "klibc/algorithm.hpp"
 #include "klibc/function.hpp"
 
 MemoryMap *sortMemoryMap(MemoryMap *mapIn)
@@ -18,5 +19,30 @@ MemoryMap *sortMemoryMap(MemoryMap *mapIn)
 	}
 
 	return mapIn;
+}
+
+bool memoryInUsableRange(
+		const MemoryMap* map,
+		uintptr_t address,
+		size_t length)
+{
+	bool inRange = false;
+	for (uint32_t i = 0; i < map->e820EntryCount; i++)
+	{
+		if (map->e820Entries[i].type != 1) continue;
+		if (!(address > map->e820Entries[i].baseAddress)) continue;
+
+		// don't bother checking if address is lower than the top of the range
+		// because if address + length is lower than the top of the range
+		// address must be contained within the range, and if address + lengths
+		// is not in the range we can't use the block anyway.
+		if (address + length <
+				map->e820Entries[i].baseAddress + map->e820Entries[i].length)
+		{
+			inRange = true;
+			break;
+		}
+	}
+	return inRange;
 }